feat: 문제 4개까지만 등록 가능하게
