+++ client/src/components/RoomSettingModal/RandomProblem.tsx
@@ -42,9 +42,13 @@ export default function RandomProblem({
         boj_problem_id: problem.bojProblemId,
         url: `https://www.acmicpc.net/problem/${problem.bojProblemId}`,
         level: problem.level,
-        tag: problem.tags.map((tag) => tag.name),
+        tag: problem.tags.map((tag:Tag) => tag.name),
       }));
-
+      // max 4 problems
+      if (problemList.length + newProblems.length > 4) {
+        alert('문제는 최대 4개까지만 등록할 수 있습니다.');
+        return;
+      }
       setProblemList([...problemList, ...newProblems]);
     } catch (error) {
       alert('문제를 가져오는데 실패했습니다.');
